+++ display.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void display_tab(float *tab, int N){
+void display_tab(double *tab, int N){
     int i,j;
     for(i = 0; i < N; i++){
         for(j = 0; j < N; j++){
@@ -11,7 +11,7 @@ void display_tab(float *tab, int N){
     }
 }
 
-void display_tab_res(float *tab, int N){
+void display_tab_res(double *tab, int N){
     int i,j;
     for(i = 0; i < N; i++){
         printf("%.20f \n",tab[i]);
