+++ testMatrice.h
@@ -1,20 +1,20 @@
 #ifndef H_TEST_MATRICE
 #define H_TEST_MATRICE
 
-void init_zero(float tab[], int N);
-void transpose(float tab[], int N);
-void a_bord1(float tab[], int N);
-void a_bord2(float tab[], int N);
-void ding_dong(float tab[], int N);
-void franc(float tab[], int N);
-void hilbertPlus(float tab[], int N, int plus);
-void hilbert(float tab[], int N);
-void hilbert2(float tab[], int N);
-void kms(float tab[], int N);
-void lehmer(float tab[], int N);
-void lotkin(float tab[], int N);
-void moler(float tab[], int N);
-void matrice_creuse(float tab[], int N, int pourcent);
-void matrice_sym_pos(float tab[], int N);
+void init_zero(double tab[], int N);
+void transpose(double tab[], int N);
+void a_bord1(double tab[], int N);
+void a_bord2(double tab[], int N);
+void ding_dong(double tab[], int N);
+void franc(double tab[], int N);
+void hilbertPlus(double tab[], int N, int plus);
+void hilbert(double tab[], int N);
+void hilbert2(double tab[], int N);
+void kms(double tab[], int N);
+void lehmer(double tab[], int N);
+void lotkin(double tab[], int N);
+void moler(double tab[], int N);
+void matrice_creuse(double tab[], int N, int pourcent);
+void matrice_sym_pos(double tab[], int N);
 
 #endif
